FEAT : BDBD-255 메인화면 만들기
