+++ BidderBidder/Assets.xcassets/Contents.json
