+++ BidderBidder/Assets.xcassets/product/Contents.json
