+++ BidderBidder/Assets.xcassets/product/airfresh.imageset/Contents.json
