+++ BidderBidder/Assets.xcassets/product/apple_pencil.imageset/Contents.json
