+++ BidderBidder/Assets.xcassets/product/apple_pencil.imageset/apple_pencil.jpg
