+++ BidderBidder/Assets.xcassets/product/blanket.imageset/Contents.json
