+++ BidderBidder/Assets.xcassets/product/blanket.imageset/blanket.jpg
