+++ BidderBidder/Assets.xcassets/product/carrot.imageset/Contents.json
