+++ BidderBidder/Assets.xcassets/product/carrot.imageset/carrot.jpg
