+++ BidderBidder/Assets.xcassets/product/chair.imageset/Contents.json
