+++ BidderBidder/Assets.xcassets/product/chair.imageset/chair.jpg
