+++ BidderBidder/Assets.xcassets/product/img2.imageset/Contents.json
