+++ BidderBidder/Assets.xcassets/product/ipad_pro.imageset/Contents.json
