+++ BidderBidder/Assets.xcassets/product/ipad_pro.imageset/ipad_pro.png
