+++ BidderBidder/Assets.xcassets/product/magic_hat.imageset/Contents.json
