+++ BidderBidder/Assets.xcassets/product/magic_hat.imageset/magic_hat.jpg
