+++ BidderBidder/Assets.xcassets/product/stand.imageset/Contents.json
