+++ BidderBidder/Assets.xcassets/product/stand.imageset/stand.jpg
