+++ BidderBidder/Assets.xcassets/product/table.imageset/table.jpg
