+++ BidderBidder/Supporting File/ColorManage.swift
